+++ src/index.js
@@ -15,14 +15,13 @@ import Request from './request';
 import { RequestList, openRequestList } from './request_list';
 import { openRequestQueue } from './request_queue';
 import SettingsRotator from './settings_rotator';
-import { apifyClient, getMemoryInfo, isAtHome, publicUtils, logSystemInfo } from './utils';
+import { apifyClient, getMemoryInfo, isAtHome, publicUtils, logSystemInfo, printOutdatedSdkWarning } from './utils';
 import { puppeteerUtils } from './puppeteer_utils';
 import { socialUtils } from './utils_social';
 import { enqueueLinks } from './enqueue_links/enqueue_links';
 import PseudoUrl from './pseudo_url';
 import LiveViewServer from './live_view/live_view_server';
-
-/* globals module */
+import { requestAsBrowser } from './utils_request';
 
 // Increase the global limit for event emitter memory leak warnings.
 EventEmitter.defaultMaxListeners = 50;
@@ -39,6 +38,9 @@ if (process.env[EMULATION_ENV_VAR]) {
 
 // Logging some basic system info (apify and apify-client version, NodeJS version, ...).
 logSystemInfo();
+
+// Log warning if SDK is outdated.
+printOutdatedSdkWarning();
 
 /**
  * The following section describes all functions and properties provided by the `apify` package,
@@ -97,6 +99,7 @@ module.exports = {
         social: socialUtils,
         log,
         enqueueLinks,
+        requestAsBrowser,
     }),
 };